move photo update to utils/profileImage.tsx
